refactor: move preKeyBundleToUserClients to PreKeyBundle module
