+++ packages/core/src/messagingProtocols/proteus/Utility/PreKeyBundle/index.ts
@@ -17,4 +17,4 @@
  *
  */
 
-export * from './PrekeyHandler';
+export * from './PreKeyBundle';
