+++ packages/core/src/messagingProtocols/proteus/Utility/Recipients.ts
@@ -21,7 +21,7 @@ import {APIClient} from '@wireapp/api-client/lib/APIClient';
 import {QualifiedUserClients, UserClients} from '@wireapp/api-client/lib/conversation';
 import {QualifiedId, QualifiedUserPreKeyBundleMap, UserPreKeyBundleMap} from '@wireapp/api-client/lib/user';
 
-import {getQualifiedPreKeyBundle, getPreKeyBundleMap} from './PreKeyBundle';
+import {getQualifiedPreKeyBundle, getPreKeyBundleMap} from './PreKeyBundle/PreKeyBundle';
 
 import {getConversationQualifiedMembers} from '../../../conversation/ConversationService/Utility/getConversationQualifiedMembers';
 import {isQualifiedUserClients, isUserClients} from '../../../util';
